fix some more 64bit warnings/issues
git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@3799 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/windows/winddraw.c
@@ -1087,7 +1087,7 @@ tryagain:
 	}
 
 	// align the destination to 16 bytes
-	dstxoffs = (((size_t)blit_desc.lpSurface + 16) & ~15) - (size_t)blit_desc.lpSurface;
+	dstxoffs = (int)((((size_t)blit_desc.lpSurface + 16) & ~15) - (size_t)blit_desc.lpSurface);
 	dstxoffs /= (dstdepth / 8);
 
 	// perform the low-level blit
